feat: moved components to different folder
