--- src/entities/ICreateAccountStateKey.ts
@@ -1,3 +1,0 @@
-import { ICreateAccountState } from "./ICreateAccountState";
-
-export type ICreateAccountStateKey = keyof ICreateAccountState;
