--- src/entities/ICreateAccountStateKeyValue.ts
@@ -1,3 +1,0 @@
-import { IOption } from "./IOption";
-
-export type ICreateAccountStateKeyValue = string | IOption | Date;
