+++ src/entities/IPersonalAccountStateKey.ts
@@ -1,0 +1,3 @@
+import { IPersonalAccountState } from "./IPersonalAccountState";
+
+export type IPersonalAccountStateKey = keyof IPersonalAccountState;
