+++ src/entities/IPersonalAccountStateKeyValue.ts
@@ -1,0 +1,3 @@
+import { IOption } from "./IOption";
+
+export type IPersonalAccountStateKeyValue = string | IOption | Date;
