+++ src/screens/ScreenCreateAccount/index.tsx
@@ -1,9 +1,9 @@
 import { VStack } from "@gluestack-ui/themed";
 import { Text } from "../../components/atoms/Text";
 import { PublicTemplateContainer } from "../../components/templates/PublicTemplateContainer";
-import { PersonalDetails } from "./components/PersonalDetails";
-import { PasswordDetails } from "./components/PasswordDetails";
-import { CompanyDetails } from "./components/CompanyDetails";
+import { PersonalDetails } from "../../components/organisms/PersonalDetails";
+import { PasswordDetails } from "../../components/organisms/PasswordDetails";
+import { CompanyDetails } from "../../components/organisms/CompanyDetails";
 import { Link } from "../../components/atoms/Link";
 import { Button } from "../../components/atoms/Button";
 import { CreateAccountProvider, useCreateAccountContext } from "./context";
